requirements.txt
# CharForgeX Requirements
# Ultra-realistic LoRA training and media generation

# Core dependencies
torch>=2.1.0
torchvision>=0.16.0
diffusers>=0.27.0
transformers>=4.38.0
accelerate>=0.27.0
safetensors>=0.4.0
omegaconf>=2.3.0
pyyaml>=6.0

# Image processing
Pillow>=10.0.0
opencv-python>=4.8.0
imageio>=2.31.0
imageio-ffmpeg>=0.4.9
scikit-image>=0.21.0
imagehash>=4.3.1

# Face detection and processing
facenet-pytorch>=2.5.3
insightface>=0.7.3
onnxruntime-gpu>=1.16.0

# Quality assessment
pyiqa>=0.1.10

# Captioning models
clip-interrogator>=0.6.0
salesforce-lavis  # For BLIP2

# LoRA training
peft>=0.8.0
bitsandbytes>=0.42.0

# ControlNet and advanced features
controlnet-aux>=0.0.7
ip-adapter

# Video generation
imageio-ffmpeg
moviepy>=1.0.3

# Audio processing (for video sync)
librosa>=0.10.1
soundfile>=0.12.1
pydub>=0.25.1

# Frame interpolation
cupy-cuda12x  # For RIFE
einops>=0.7.0

# Cloud integration
runpod>=1.5.0
boto3>=1.34.0  # For S3-compatible storage

# API and web
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
gradio>=4.16.0
websockets>=12.0

# CLI and utilities
click>=8.1.7
tqdm>=4.66.0
rich>=13.7.0
python-dotenv>=1.0.0

# Hugging Face
huggingface-hub[hf-transfer]>=0.20.0

# Optional but recommended
xformers>=0.0.23  # Memory-efficient attention
triton>=2.1.0  # torch.compile support

# Development
pytest>=7.4.0
black>=23.12.0
